# [TOPI] Update softmax compute and CPU schedule