--- a/topi/python/topi/cuda/softmax.py
+++ b/topi/python/topi/cuda/softmax.py
@@ -38,8 +38,9 @@ def schedule_softmax(outs):
     outs = [outs] if isinstance(outs, tvm.tensor.Tensor) else outs
     s = tvm.create_schedule([x.op for x in outs])
     softmax = outs[0]
-    max_elem = softmax.op.input_tensors[1]
-    expsum = softmax.op.input_tensors[2]
+    exp = softmax.op.input_tensors[0]
+    expsum = softmax.op.input_tensors[1]
+    maxelem = s[exp].op.input_tensors[1]
 
     if len(softmax.shape) > 2:
         for op in [max_elem.op, expsum.op, softmax.op]: