--- a/topi/python/topi/opengl/softmax.py
+++ b/topi/python/topi/opengl/softmax.py
@@ -44,11 +44,15 @@ def schedule_softmax(outs):
         exp = softmax.op.input_tensors[0]
         max_elem = s[exp].op.input_tensors[1]
     elif op_tag == 'log_softmax_output':
+        exp = None
         max_elem = softmax.op.input_tensors[1]
         expsum = softmax.op.input_tensors[2]
     else:
         raise ValueError('Tag is expected to be softmax_output or log_softmax_output. \
                          Got {0}'.format(op_tag))
+
+    if exp != None:
+        s[exp].opengl()
 
     s[max_elem].opengl()
     s[expsum].opengl()