This change improves performance for softmax by simplifying the computation and writing a schedule that supports better parallelization.

**Compute**: Currently, `exp(input - max)` is computed twice: once in the `_compute_expsum` stage and once in the `_normalize` stage. This change adds an extra stage to compute this tensor once. It is then re-used in the `_compute_expsum` and `_normalize` stages.

**Schedule**: Currently, the schedule only parallelizes the `_normalize` stage of the computation. This change puts all stages of computation under a common root and parallelizes the outer dimensions.

The following results are with a tensor of shape `(1,12,128,128)` and `axis=-1`. This simulates the softmax in BERT base. The CPU is Intel Xeon E5-2650, and the Relay target string is `llvm -mcpu=core-avx2`.

| TVM_NUM_THREADS  | Latency in ms (master branch) | Latency in ms (new branch) |
| ------------- | ------------- | ------------- |
| 1 | 4.7 | 3.0 |
| 2 | 3.8 | 1.8 |
| 4 | 3.3 | 1.0 |
| 8 | 3.1 | 0.74 |
| 16 | 3.2 | 0.55 |