@tqchen 2019-08-01
Thank you @soiferj , can you check the CI problem?