@soiferj 2019-08-01
Yeah, I'm taking a look at the CI failure now. It seems to be an issue in the CUDA schedule. I will work on it.