@soiferj 2019-08-02
The CI issue is fixed.