@anijain2305 2020-02-13
Another suggestion - https://discuss.tvm.ai/t/softmax-sequence-of-relay-ops/5686